server/requirements.txt
gevent==1.4.0
Werkzeug==0.14.1
Flask-Script==2.0.6
Flask==1.0.2
Flask-Cache==0.13.1
Flask-Docs==0.3.9
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.7.0
Flask-Script==2.0.6
